# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Alphabet AI - Adaptive K-12 ELA Mastery Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the returnTo path on success or /api/login on failure.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'

  # ── STUDENT PROFILE ────────────────────────────────────────────────────────
...
  /placement/start:
    post:
      operationId: startPlacement
      tags: [placement]
      summary: Start a new adaptive placement session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlacementStartInput"
      responses:
        "201":
          description: New placement session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlacementSession"
...
  /placement/{sessionId}/answer:
    post:
      operationId: submitPlacementAnswer
      tags: [placement]
      summary: Submit answer for placement question
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlacementAnswer"
      responses:
        "200":
          description: Answer result with updated theta
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlacementAnswerResult"
  /placement/{sessionId}/result:
    get:
      operationId: getPlacementResult
      tags: [placement]
      summary: Get placement result after completion
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Placement result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlacementResult"

  # ── DAILY PRACTICE ─────────────────────────────────────────────────────────
...
  /llm/generate-question:
    post:
      operationId: generateQuestion
      tags: [llm]
      summary: Generate an adaptive practice question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionGenerateInput"
      responses:
        "200":
          description: Generated question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdaptiveQuestion"
...
components:
  schemas:
...
    PlacementStartInput:
      type: object
      properties:
        startTheta:
          type: number
    PlacementSession:
      type: object
      required: [id, studentId, status, questionCount, theta]
      properties:
        id:
          type: string
        studentId:
          type: string
        status:
          type: string
          enum: [in_progress, completed]
        questionCount:
          type: integer
        theta:
          type: number
        thetaSe:
          type: number
        createdAt:
          type: string
    AdaptiveQuestion:
      type: object
      required: [id, skillCode, skillName, questionText, activityType, options]
      properties:
        id:
          type: string
        skillCode:
          type: string
        skillName:
          type: string
        domain:
          type: string
        questionText:
          type: string
        passage:
          type: ["string", "null"]
        activityType:
          type: string
          enum: [listen_repeat, see_tap, say_it, write_it, read_it, multiple_choice]
        options:
          type: array
          items:
            $ref: "#/components/schemas/QuestionOption"
        correctOptionId:
          type: string
        explanation:
          type: ["string", "null"]
        difficulty:
          type: number
    QuestionOption:
      type: object
      required: [id, text]
      properties:
        id:
          type: string
        text:
          type: string
    PlacementAnswer:
      type: object
      required: [questionId, selectedOptionId, skillCode, correct]
      properties:
        questionId:
          type: string
        selectedOptionId:
          type: string
        skillCode:
          type: string
        correct:
          type: boolean
        timeSpentSeconds:
          type: integer
    PlacementAnswerResult:
      type: object
      required: [correct, newTheta, thetaSe, complete]
      properties:
        correct:
          type: boolean
        newTheta:
          type: number
        thetaSe:
          type: number
        complete:
          type: boolean
        explanation:
          type: ["string", "null"]
    PlacementResult:
      type: object
      required: [sessionId, diagnosedGradeLevel, placementPathway, thetaFinal]
      properties:
        sessionId:
          type: string
        diagnosedGradeLevel:
          type: string
        placementPathway:
          type: string
          enum: [foundation, developing, proficient, advanced]
        thetaFinal:
          type: number
        accuracyPct:
          type: number
        strandStrengths:
          type: array
          items:
            type: string
        strandGaps:
          type: array
          items:
            type: string

    # ── PRACTICE ──────────────────────────────────────────────────────────────
...
    QuestionGenerateInput:
      type: object
      required: [skillCode, gradeLevel, activityType]
      properties:
        skillCode:
          type: string
        gradeLevel:
          type: string
        activityType:
          type: string
        interests:
          type: array
          items:
            type: string
        culturalContext:
          type: array
          items:
            type: string
        homeLanguage:
          type: string
        theta:
          type: number
...
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: ["string", "null"]
          format: email
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        profileImageUrl:
          type: ["string", "null"]
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── ANALYTICS ─────────────────────────────────────────────────────────────
...
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — Bearer <sid>.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string