# lib/api-zod/src/generated/types/placementStartInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Alphabet AI - Adaptive K-12 ELA Mastery Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PlacementStartInput {
  startTheta?: number;
}